scraper/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = ""
authors = [
    {name = "Christopher Bonner",email = "cbonner.dev@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright (>=1.50.0,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "ics (>=0.7.2,<0.8.0)"
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "--import-mode=importlib"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
